requirements.txt
wechaty
wechaty-puppet
wechaty-puppet-service
wechaty-plugin-contrib
apscheduler
pandas
openpyxl
pyparsing
dataclasses-json
python-dotenv
pdfkit
asq
typed-argument-parser
xlrd==1.2.0